# Reduces lock-contention in high-performance usage of go-metrics.