Remove lock contention in go-metrics in a few key places. For all of these changes there is expectation around CPU performance gain but the **real benefit is reduced lock contention which shows up as significant bottlenecks in both Go 1.9 and Go 1.10** for high performance services.

1. **gauge_float64.go**
- Removes the mutex lock around the internal value field in favor of atomic ops.
- Changes the value field to be an uint64 (no precision is lost) and does conversions
- Utilizes atomic.Load/atomic.Store to reduce lock contention
- Adds a parallel benchmark

```sh
BEFORE:
BenchmarkGuageFloat64Parallel-8   	20000000	        88.1 ns/op	       0 B/op	       0 allocs/op
AFTER:
BenchmarkGuageFloat64Parallel-8   	50000000	        28.8 ns/op	       0 B/op	       0 allocs/op
```

2. **ewma.go**
- Removes lock contention around a few fields in favor of atomic ops.
- For the mutex lock in place this lock is only taken on the first Tick() operation only
- Adds a parallel benchmark

```sh
BEFORE:
BenchmarkEWMAParallel-8   	20000000	        99.4 ns/op	       0 B/op	       0 allocs/op
AFTER:
BenchmarkEWMAParallel-8   	20000000	        74.5 ns/op	       0 B/op	       0 allocs/op
```

3. **meter.go**
- Removes lock contention around multiple fields in favor of atomic ops.
- Mutex locking is left in places where there is not high concurrency such as tickMeters and Stop.
- Adds a parallel benchmark

```sh
BEFORE:
BenchmarkMeterParallel-8   	 3000000	       467 ns/op	       0 B/op	       0 allocs/op
AFTER:
BenchmarkMeterParallel-8   	10000000	       169 ns/op	       0 B/op	       0 allocs/op
```