@mihasya 2018-04-05
Howdy, thanks for the PR!

I don't have a lot of visibility into golang version adoption. Do you know of any data out there about how many folks might still be on 1.2? I'm reticent to drop support for a version seemingly in order to enable a parallel benchmark without knowing the numbers.