@deckarep 2018-04-05
@mihasya - I don't have numbers...but the unfortunate thing is the 1.2 doesn't allow us to leverage the Parallel Benchmark features...alternatively we can simply remove it...or perhaps there is a build flag to not build that part of the code if we're on 1.2.

I would say though, most if not all large projects with many contributors with Go don't support anything further then 1.5 these days...so 1.2 is really, really behind.

Additionally, this code change doesn't actually prevent anyone from running on 1.2...it just prevents the parallel benchmarks from being compiled.