@deckarep 2018-04-05
One extra note: I'm adding a few more changes right now..so feel free to not review this just yet.