@mihasya 2018-04-05
Oh great point, it's in a `_test` file, so we'd just not be testing on it. That makes me feel better.

(to be clear, intuitively it seems fine to drop a version of Go from 2012, but I'm just not plugged into the scene enough to know how "up to date" folks are)