@deckarep 2018-04-05
@mihasya - i totally understand. Yes, 1.2 is long considered dead and highly not recommended for anyone still on it due to security updates.