@deckarep 2018-04-06
@mihasya - let me know if you're fine with the PR as it is or if you want me to revert the .travis.yaml change to put Go 1.2 back?  I think losing the Parallel Bench unit-tests is fine for us so long as we can move forward with merging this change.