@mihasya 2018-04-06
@deckarep no no, I'm totally fine with dropping Go 1.2, just haven't had a chance to look at the more recent changes.

The biggest thing I want to double check is that we have good tests for validating concurrent behavior. It's been a little while since I've looked at the tests we do have in earnest, and I haven't had a chance to carve out some time to review the coverage. Anything that changes how things are calculated makes me a little nervous - messing up metrics output is a super quick way to pitchforks :)

I would say this has a very high chance of getting merged as is because it "looks good" to me, so it would be safe to proceed using this fork if you're comfortable with it and want to move faster internally than I can merge the PR.