@mihasya 2018-04-06
@deckarep due to an unexpected sick day (BOOOO!), I've spent some time exercising my Go muscles and adding some very basic concurrency tests to all the things that have changed. It looks like there's a race in the Meter implementation.

I've cherry-picked the same test onto master, and it doesn't appear to fail. I'm probably going to take a break and nap, but may take a look at fixing the race later if you don't beat me to it.