@deckarep 2018-04-06
@mihasya - ok, this race was my bad...i forgot to port over one method that has updated behavior on the *stop* functionality. 

I pushed up a branch just now and handled the Mark method appropriately.