@deckarep 2018-04-06
Thank you so much, I totally agree...nailing down that race was most excellent.
Hope you feel better and thanks for being an awesome open-source maintainer.

Cheers @mihasya!